Removed logic used determine if user has dismissed a previous touchpoint notification or not
When user traverse through touchpoints, notifications will be stacked in the same notification window.  Even if the user has previously seen the touchpoint, if he comes back after the keepalive time, the same message will appear again, stacked onto the previous one.  The only way to clear old notifications is to have the user click on it, which brings up the CardListActivity and clears the notification events queue.